Setup redis caching boilerplate, move gateway client into spring bean

+++ core/src/main/kotlin/org/dreamexposure/discal/core/object/network/discal/BotInstanceData.kt
@@ -29,7 +29,7 @@ data class BotInstanceData private constructor(
                     .map { guildCount ->
                         BotInstanceData(
                                 instanceData = InstanceData(),
-                                shardIndex = Application.getShardIndex().toInt(),
+                                shardIndex = Application.getShardIndex(),
                                 shardCount = Application.getShardCount(),
                                 guilds = guildCount
                         )
